requirements.txt
annotated-types==0.7.0
astroid==3.2.4
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dill==0.3.8
Flask==3.0.3
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.144.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.62.3
httplib2==0.22.0
idna==3.8
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
lxml==5.3.0
Markdown==3.7
MarkupSafe==2.1.5
mccabe==0.7.0
numpy==2.1.1
oauthlib==3.2.2
pandas==2.2.2
platformdirs==4.3.2
praw==7.7.1
prawcore==2.4.0
proto-plus==1.24.0
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.9.1
pydantic_core==2.23.3
pylint==3.2.7
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrends==4.9.2
pytz==2024.2
regex==2024.9.11
requests==2.32.3
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
tenacity==9.0.0
tomli==2.0.1
tomlkit==0.13.2
tqdm==4.66.5
tweepy==4.14.0
typing_extensions==4.12.2
tzdata==2024.1
update-checker==0.18.0
uritemplate==4.1.1
urllib3==2.2.2
websocket-client==1.8.0
Werkzeug==3.0.4